requirements.txt
discord.py
asyncpraw
pillow==6.2.1
pandas
opencv-python-headless
async_google_trans_new
discord-ext-paginator
bs4
httpx
lxml
typing
pyquery
discord-pretty-help